kafka events refactoring based on comments from code review
