+++ galasa-parent/dev.galasa.framework/src/main/java/dev/galasa/framework/ITestRunnerDataProvider.java
@@ -7,7 +7,7 @@ package dev.galasa.framework;
 
 import java.util.Properties;
 
-import dev.galasa.framework.internal.events.IEventsPublisher;
+import dev.galasa.framework.internal.runner.IFrameworkEventsProducer;
 import dev.galasa.framework.spi.IConfigurationPropertyStoreService;
 import dev.galasa.framework.spi.IDynamicStatusStoreService;
 import dev.galasa.framework.spi.IResultArchiveStore;
@@ -26,5 +26,5 @@ public interface ITestRunnerDataProvider {
     public Properties getOverrideProperties();
     public ITestRunManagers createTestRunManagers(GalasaTest galasaTest) throws TestRunException;
     public IFileSystem getFileSystem();
-    public IEventsPublisher getEventsPublisher();
+    public IFrameworkEventsProducer getEventsProducer();
 }
